perf👌: 修改os内核

+++ source/required/utility/os/os_kernel.h
@@ -75,7 +75,7 @@ static inline i32 kernel_resume_(void)
  * @return 结果
  * @retval 0 成功
  */
-static inline long kernel_suspend_(void)
+static inline i32 kernel_suspend_(void)
 {
         #if ((FREERTOS == 1) && (FREERTOS_TASK == 1))
         vTaskSuspendAll();
